Pop Up Admin

+++ admin.php
@@ -109,10 +109,13 @@ if (isset($_POST["btnSearchAdmin"])) {
     header("Location: #collections");
 }
 
+$showModal = "";
+
 if (isset($_POST["submitBrand"])) {
     $tempNameBrand = $_POST["addBrand"];
     if ($tempNameBrand == "") {
-        alert("Add Brand Failed!");
+        $showModal = "brandFail";
+        // alert("Add Brand Failed!");
     } else {
         $bikinId = "BRN";
         $tempBrand = query("SELECT * FROM brand");
@@ -133,15 +136,17 @@ if (isset($_POST["submitBrand"])) {
             "name_brand" => $tempNameBrand
         ];
         insertBrand($brandBaru);
-        alert("Add Brand Success!");
+        $showModal = "brandSuccess";
+        // alert("Add Brand Success!");
     }
-    $_SESSION["data"] = "produk";
-    echo "<script> document.location.href = 'admin.php'; </script>";
+    // $_SESSION["data"] = "produk";
+    // echo "<script> document.location.href = 'admin.php'; </script>";
 }
 if (isset($_POST["submitKategori"])) {
     $tempNameKategori = $_POST["addKategori"];
     if ($tempNameKategori == "") {
-        alert("Add Categories Failed!");
+        $showModal = "categoryFail";
+        // alert("Add Categories Failed!");
     } else {
         $bikinId = "KTG";
         $tempKategori = query("SELECT * FROM kategori");
@@ -162,10 +167,11 @@ if (isset($_POST["submitKategori"])) {
             "name_kategori" => $tempNameKategori
         ];
         insertKategori($KategoriBaru);
-        alert("Add Categories Success!");
+        $showModal = "categorySuccess";
+        // alert("Add Categories Success!");
     }
-    $_SESSION["data"] = "produk";
-    echo "<script> document.location.href = 'admin.php'; </script>";
+    // $_SESSION["data"] = "produk";
+    // echo "<script> document.location.href = 'admin.php'; </script>";
 }
 
 if (isset($_POST["edit"])) {
@@ -192,15 +198,17 @@ if (isset($_POST["activate"])) {
         $query = "UPDATE produk SET status_produk = '1' WHERE id_produk = '$id'";
         mysqli_query($conn, $query);
         $_SESSION["dataAdmin"] = query("SELECT * FROM produk");
-        alert("Activate Product Success!");
+        $showModal = "activeProduct";
+        // alert("Activate Product Success!");
     } else if ($_SESSION["data"] == "user") {
         $query = "UPDATE user SET status_user = '1' WHERE id_user = '$id'";
         mysqli_query($conn, $query);
         $_SESSION["dataAdmin"] = query("SELECT * FROM user");
-        alert("Activate User Success!");
+        $showModal = "activeUser";
+        // alert("Activate User Success!");
     }
 
-    echo "<script> document.location.href = 'admin.php#collections'; </script>";
+    // echo "<script> document.location.href = 'admin.php#collections'; </script>";
 }
 if (isset($_POST["deactivate"])) {
     $id = $_POST["idData"];
@@ -219,20 +227,23 @@ if (isset($_POST["deactivate"])) {
         $query = "UPDATE produk SET status_produk = '0' WHERE id_produk = '$id'";
         mysqli_query($conn, $query);
         $_SESSION["dataAdmin"] = query("SELECT * FROM produk");
-        alert("Deactivate Product Success!");
+        $showModal = "deactivateProduct";
+        // alert("Deactivate Product Success!");
     } else if ($_SESSION["data"] == "user") {
         $query = "UPDATE user SET status_user = '0' WHERE id_user = '$id'";
         mysqli_query($conn, $query);
         $_SESSION["dataAdmin"] = query("SELECT * FROM user");
-        alert("Deactivate User Success!");
+        $showModal = "deactivateUser";
+        // alert("Deactivate User Success!");
     }
 
 
-    echo "<script> document.location.href = 'admin.php#collections'; </script>";
+    // echo "<script> document.location.href = 'admin.php#collections'; </script>";
 }
 if (isset($_POST["submitProduk"])) {
     $safe = true;
 
+    $error = [];
     $insertData = [];
     for ($i = 0; $i < $_SESSION["jumlahInput"]; $i++) {
         $nameProduk = $_POST["addName" . $i];
@@ -244,6 +255,8 @@ if (isset($_POST["submitProduk"])) {
 
         if ($nameProduk == "" || $brandProduk == "" || $categoryProduk == "" || $stockProduk == "" || $priceProduk == "" || $descriptionProduk == "") {
             $safe = false;
+            $err = "All Fields Must Be Filled!";
+            array_push($error, $err);
         } else {
             $target_dir = "asset/product/";
             $target_file = $target_dir . basename($_FILES["addImage" . $i]["name"]);
@@ -256,25 +269,33 @@ if (isset($_POST["submitProduk"])) {
                 $uploadOk = true;
             } else {
                 $uploadOk = false;
-                alert("File is not an image.");
+                $err = "File is not an image.";
+                array_push($error, $err);
+                // alert("File is not an image.");
             }
 
             // Check if file already exists
             if (file_exists($target_file) && $uploadOk == true) {
                 $uploadOk = false;
-                alert("Sorry, file already exists.");
+                $err = "File already exists.";
+                array_push($error, $err);
+                // alert("Sorry, file already exists.");
             }
 
             // Check file size
             if ($_FILES["addImage" . $i]["size"] > 10000000 && $uploadOk == true) {
                 $uploadOk = false;
-                alert("Sorry, your file is too large.");
+                $err = "File is too large.";
+                array_push($error, $err);
+                // alert("Sorry, your file is too large.");
             }
 
             // Allow certain file formats
             if ($imageFileType != "jpg" && $uploadOk == true) {
                 $uploadOk = false;
-                alert("Sorry, only JPG files are allowed.");
+                $err = "Only JPG files are allowed.";
+                array_push($error, $err);
+                // alert("Sorry, only JPG files are allowed.");
             }
 
             // Check if $uploadOk is set to 0 by an error
@@ -284,12 +305,15 @@ if (isset($_POST["submitProduk"])) {
                 if (move_uploaded_file($_FILES["addImage" . $i]["tmp_name"], $target_file)) {
                     $safe = true;
                 } else {
-                    alert("Sorry, there was an error uploading your file.");
+                    $err = "There was an error uploading your file.";
+                    array_push($error, $err);
+                    // alert("Sorry, there was an error uploading your file.");
                     $safe = false;
                 }
             }
         }
 
+        $path = "";
         if ($safe) {
             $path = $target_dir . htmlspecialchars(basename($_FILES["addImage" . $i]["name"]));
 
@@ -325,14 +349,18 @@ if (isset($_POST["submitProduk"])) {
     if ($safe) {
         insertProduk($insertData);
 
-        alert("Add Product Success!");
+        $showModal = "productSuccess";
+        // alert("Add Product Success!");
     } else {
-        alert("Add Product Failed!");
-        unlink($path);
+        $showModal = "productFail";
+        // alert("Add Product Failed!");
+        if ($path != "") {
+            unlink($path);
+        }
     }
     $_SESSION["data"] = "produk";
     $_SESSION["dataAdmin"] = query("SELECT * FROM produk");
-    echo "<script> document.location.href = 'admin.php'; </script>";
+    // echo "<script> document.location.href = 'admin.php'; </script>";
 }
 
 // PAGINATION
@@ -653,7 +681,7 @@ if (isset($_POST["go"])) {
                             }
                             ?>
                         </select>
-                        <button type="submit" class="btn btn-outline-dark fs-5 d-flex align-items-center me-3 text-center" style="width: 6vw; height: 5vh;" name="choose">Choose</button>
+                        <button type="submit" class="btn btn-outline-dark fs-5 d-flex align-items-center me-3 justify-content-center" style="width: 6vw; height: 5vh;" name="choose">Choose</button>
                         <button type="submit" class="btn btn-outline-danger fs-5 d-flex align-items-center justify-content-center" style="width: 6vw; height: 5vh;" name="report">Report</button>
                     </div>
                 </div>
@@ -1383,6 +1411,178 @@ if (isset($_POST["go"])) {
         ?>
     </div>
 
+    <!-- Modal BrandFail -->
+    <div class="modal fade" id="modalBrandFail" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Brand Failed!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal BrandSuccess -->
+    <div class="modal fade" id="modalBrandSuccess" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Brand Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal CategoryFail -->
+    <div class="modal fade" id="modalCategoryFail" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Categories Failed!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal CategorySuccess -->
+    <div class="modal fade" id="modalCategorySuccess" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Categories Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal ActiveProduct -->
+    <div class="modal fade" id="modalActiveProduct" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Activate Product Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal ActiveUser -->
+    <div class="modal fade" id="modalActiveUser" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Activate User Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal DeactivateProduct -->
+    <div class="modal fade" id="modalDeactivateProduct" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Deactivate Product Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal DeactivateUser -->
+    <div class="modal fade" id="modalDeactivateUser" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Deactivate User Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal ProductFail -->
+    <div class="modal fade" id="modalProductFail" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Product Failed!</h4>
+                        <div class="errorDetail">
+                            <span>Error:</span><br>
+                            <?php
+                            for ($i = 0; $i < count($error); $i++) {
+                                $strError = $error[$i];
+                            ?>
+                                <span>- <?= $strError ?></span><br>
+                            <?php
+                            }
+                            ?>
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+    <!-- Modal ProductSuccess -->
+    <div class="modal fade" id="modalProductSuccess" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered">
+            <div class="modal-content">
+                <div class="modal-body">
+                    <div class="tengah p-4">
+                        <h4 class="text-center">Add Product Success!</h4>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-none border-0 w-100 fw-bold fs-5" data-bs-dismiss="modal">OK</button>
+
+                </div>
+            </div>
+        </div>
+    </div>
+
     <script>
         function clearForm() {
             document.getElementsById("addName").value = "";
@@ -1392,8 +1592,94 @@ if (isset($_POST["go"])) {
         }
     </script>
 </body>
+<script script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
+<script src="jquery-3.6.1.min.js"></script>
+<script>
+    // $('#btnTest').click(function() {
+    //     $('#testingModal').modal('show');
+    // });
+</script>
+<?php
+if ($showModal == "brandFail") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalBrandFail").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "brandSuccess") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalBrandSuccess").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "categoryFail") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalCategoryFail").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "categorySuccess") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalCategorySuccess").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "activeProduct") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalActiveProduct").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "activeUser") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalActiveUser").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "deactivateProduct") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalDeactivateProduct").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "deactivateUser") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalDeactivateUser").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "productFail") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalProductFail").modal("show");
+    	});
+    </script>';
+}
+if ($showModal == "productSuccess") {
+    echo
+    '<script type="text/javascript">
+        $(document).ready(function(){
+    		$("#modalProductSuccess").modal("show");
+    	});
+    </script>';
+}
+?>
 
 </html>
-
-<!-- p -->
-<!-- p2 -->
